renaming from adaptor -> adapter. adding docs
