+++ src/main/java/smile/identity/core/SmileIdentityService.java
@@ -10,8 +10,8 @@ import retrofit2.Call;
 import retrofit2.Response;
 import retrofit2.Retrofit;
 import retrofit2.converter.moshi.MoshiConverterFactory;
-import smile.identity.core.adaptors.JobTypeAdaptor;
-import smile.identity.core.adaptors.PartnerParamsAdaptor;
+import smile.identity.core.adapters.JobTypeAdaptor;
+import smile.identity.core.adapters.PartnerParamsAdapter;
 import smile.identity.core.exceptions.JobFailed;
 import smile.identity.core.models.*;
 
@@ -30,7 +30,7 @@ public class SmileIdentityService {
 
         Moshi moshi = new Moshi.Builder()
                 .add(factory)
-                .add(new PartnerParamsAdaptor())
+                .add(new PartnerParamsAdapter())
                 .add(new JobTypeAdaptor())
                 .build();
